Update to react-dsfr 1.10
