+++ .storybook/manager-head.html
@@ -6,110 +6,10 @@
     href="%PUBLIC_URL%/dsfr/favicon/manifest.webmanifest"
     crossorigin="use-credentials"
 />
+<link rel="stylesheet" href="%PUBLIC_URL%/dsfr/fonts/index.css" />
 <!-- Meta tags generated by metatags.io -->
 <!-- Primary Meta Tags -->
 <title>Lunatic DSFR</title>
 <meta name="title" content="lunatic-dsfr components" />
 <meta name="description" content="@inseefr/lunatic-dsfr components" />
 
-<style>
-    @font-face {
-        font-family: Marianne;
-        src:
-            url("./dsfr/fonts/Marianne-Light.woff2") format("woff2"),
-            url("./dsfr/fonts/Marianne-Light.woff") format("woff");
-        font-weight: 300;
-        font-style: normal;
-        font-display: swap;
-    }
-
-    @font-face {
-        font-family: Marianne;
-        src:
-            url("./dsfr/fonts/Marianne-Light_Italic.woff2") format("woff2"),
-            url("./dsfr/fonts/Marianne-Light_Italic.woff") format("woff");
-        font-weight: 300;
-        font-style: italic;
-        font-display: swap;
-    }
-
-    @font-face {
-        font-family: Marianne;
-        src:
-            url("./dsfr/fonts/Marianne-Regular.woff2") format("woff2"),
-            url("./dsfr/fonts/Marianne-Regular.woff") format("woff");
-        font-weight: 400;
-        font-style: normal;
-        font-display: swap;
-    }
-
-    @font-face {
-        font-family: Marianne;
-        src:
-            url("./dsfr/fonts/Marianne-Regular_Italic.woff2") format("woff2"),
-            url("./dsfr/fonts/Marianne-Regular_Italic.woff") format("woff");
-        font-weight: 400;
-        font-style: italic;
-        font-display: swap;
-    }
-
-    @font-face {
-        font-family: Marianne;
-        src:
-            url("./dsfr/fonts/Marianne-Medium.woff2") format("woff2"),
-            url("./dsfr/fonts/Marianne-Medium.woff") format("woff");
-        font-weight: 500;
-        font-style: normal;
-        font-display: swap;
-    }
-
-    @font-face {
-        font-family: Marianne;
-        src:
-            url("./dsfr/fonts/Marianne-Medium_Italic.woff2") format("woff2"),
-            url("./dsfr/fonts/Marianne-Medium_Italic.woff") format("woff");
-        font-weight: 500;
-        font-style: italic;
-        font-display: swap;
-    }
-
-    @font-face {
-        font-family: Marianne;
-        src:
-            url("./dsfr/fonts/Marianne-Bold.woff2") format("woff2"),
-            url("./dsfr/fonts/Marianne-Bold.woff") format("woff");
-        font-weight: 700;
-        font-style: normal;
-        font-display: swap;
-    }
-
-    @font-face {
-        font-family: Marianne;
-        src:
-            url("./dsfr/fonts/Marianne-Bold_Italic.woff2") format("woff2"),
-            url("./dsfr/fonts/Marianne-Bold_Italic.woff") format("woff");
-        font-weight: 700;
-        font-style: italic;
-        font-display: swap;
-    }
-
-    @font-face {
-        font-family: Spectral;
-        src:
-            url("./dsfr/fonts/Spectral-Regular.woff2") format("woff2"),
-            url("./dsfr/fonts/Spectral-Regular.woff") format("woff");
-        font-weight: 400;
-        font-style: normal;
-        font-display: swap;
-    }
-
-    @font-face {
-        font-family: Spectral;
-        src:
-            url("./dsfr/fonts/Spectral-ExtraBold.woff2") format("woff2"),
-            url("./dsfr/fonts/Spectral-ExtraBold.woff") format("woff");
-        font-weight: 900;
-        font-style: normal;
-        font-display: swap;
-    }
-</style>
